Added checks if Pillow library is installed
